+++ docs/search/classes_2.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['defaultactivityinstancegenerator',['DefaultActivityInstanceGenerator',['../classBristolSU_1_1Support_1_1ActivityInstance_1_1DefaultActivityInstanceGenerator.html',1,'BristolSU\Support\ActivityInstance\DefaultActivityInstanceGenerator'],['../interfaceBristolSU_1_1Support_1_1ActivityInstance_1_1Contracts_1_1DefaultActivityInstanceGenerator.html',1,'BristolSU\Support\ActivityInstance\Contracts\DefaultActivityInstanceGenerator']]]
+  ['defaultactivityinstancegenerator',['DefaultActivityInstanceGenerator',['../interfaceBristolSU_1_1Support_1_1ActivityInstance_1_1Contracts_1_1DefaultActivityInstanceGenerator.html',1,'BristolSU\Support\ActivityInstance\Contracts\DefaultActivityInstanceGenerator'],['../classBristolSU_1_1Support_1_1ActivityInstance_1_1DefaultActivityInstanceGenerator.html',1,'BristolSU\Support\ActivityInstance\DefaultActivityInstanceGenerator']]]
 ];