+++ docs/search/files_6.js
@@ -2,6 +2,7 @@ var searchData=
 [
   ['handlesauthentication_2ephp',['HandlesAuthentication.php',['../HandlesAuthentication_8php.html',1,'']]],
   ['handlesauthorization_2ephp',['HandlesAuthorization.php',['../HandlesAuthorization_8php.html',1,'']]],
+  ['hashistory_2ephp',['HasHistory.php',['../HasHistory_8php.html',1,'']]],
   ['hasresource_2ephp',['HasResource.php',['../HasResource_8php.html',1,'']]],
   ['hasrevisions_2ephp',['HasRevisions.php',['../HasRevisions_8php.html',1,'']]],
   ['helpers_2ephp',['helpers.php',['../helpers_8php.html',1,'']]],